backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pypdf
pymupdf
pillow
pyodbc

openai>=1.40.0

# RAG – volitelné (nebude-li k dispozici, náš kód běží i bez FAISS)
numpy
tiktoken
faiss-cpu

# OCR stack
pdf2image
pytesseract
Pillow

numpy